ui (rename *.tsx): clarify the role of components
