+++ src/ui/readme.md
@@ -1,0 +1,9 @@
+## Cellexp UI
+
+### Editor
+
+This is the "Rule Picker" component.
+
+### Control
+
+This is the "Controllers" component.
